Integrate in APIs into Next app
