+++ README.md
@@ -12,6 +12,10 @@ This repo is a living version of the dashboard, as it evolves into an independen
 
 This application is built on the React application framework [Next.js](https://nextjs.org/).
 
+### Pre-requisites
+
+In order to use run this project locally, you will need to have [node](https://nodejs.org/en/download/) and [docker](https://docs.docker.com/desktop/) installed.
+
 ### Installation
 
 First install the project dependencies with:
@@ -22,12 +26,34 @@ npm i
 
 ### Running Locally
 
+#### Run the Database
+
+This repo comes with a development MySQL database that runs in docker based on the latest production dump (in mysql/dump.sql). You can generally start the database and just leave it running in the background.
+
+```
+npm run start-dev-db
+```
+
+If you need to refresh the database for any reason, stop and re-start it:
+
+```
+npm run stop-dev-db
+npm run start-dev-db
+```
+
+#### Running the Development Server
+
 To run, start the development server with the following command:
 
 ```
 npm run dev
 ```
 
-### Deployment
+### Running in Production
 
-// TODO 
+To run in production, first build the production assets and then start the production server:
+
+```
+npm run build
+npm run start
+```
