--- api/api.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <title>policetracker.link</title>
-
-    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.1/css/bootstrap.min.css">
-    <style>
-        body {
-            padding-top: 50px;
-        }
-    </style>
-</head>
-
-<body>
-
-    <div class="container">
-        <div class="jumbotron">
-            <h1>APIs available for pulling police shootings and police brutality</h1>
-
-            <h3> /shootings will return all records in our DB</h3>
-            <h3> /shootings/{id} will return just record number id</h3>
-            <h3> /shootings/{id1}-{id2} will return records from id1 to id2</h3>
-            <h3> /shootings/last20 will return the last (most recent) 20 records</h3>
-            <h3> /count will return the count of all shootings and brutality records</h3>
-            <h3> /count/state/name will return the count of all shootings and brutality records by state</h3>
-            <h3> /count/state/abbv will return the count of all shootings and brutality records by state abbreviation
-            </h3>
-
-            <h3> /count/shootings will return the count of all shootings records</h3>
-            <h3> /count/shootings/state/name will return the count of all shootings records by state</h3>
-            <h3> /count/shootings/state/abbv will return the count of all shootings records by state abbreviation</h3>
-            <h3> /count/shootings/statecounty will return the count of all shootings records by state and county</h3>
-
-            <h3> /count/brutality will return the count of all brutality records</h3>
-            <h3> /count/brutality/state/name will return the count of all brutality records by state</h3>
-            <h3> /count/brutality/state/abbv will return the count of all brutality records by state abbreviation</h3>
-            <h3> /count/brutality/statecounty will return the count of all brutality records by state and county</h3>
-
-
-            <h2>Future functionality will also include police brutality data</h2>
-
-        </div>
-    </div>
-
-</body>
-
-</html>
