--- api/faq.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <title>policetracker.link</title>
-
-    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.1/css/bootstrap.min.css">
-    <style>
-        body {
-            padding-top: 50px;
-        }
-    </style>
-</head>
-
-<body>
-
-    <div class="container">
-        <div class="jumbotron">
-            <h1>FAQ</h1>
-
-            <h3>Where did you get this data?</h3>
-            <h2>Data was collected by volunteers of the Grassroot Law Project</h2>
-
-            <h3>How accurate is this data?</h3>
-            <h2>Data will refined by volunteers manually at this point to remove duplicates and validate entries</h2>
-            <h2>Data is free to use, but we offer no guarantees on the data at this point in time.</h2>
-
-            <h3>How can I help?</h3>
-            <h2>Check with the Grassroot Law Project</h2>
-        </div>
-    </div>
-
-</body>
-
-</html>
