--- api/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "policetracker",
-  "version": "1.0.0",
-  "description": "tracking police violence",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "express": "^4.17.1",
-    "mysql": "^2.18.1"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.4"
-  }
-}
