+++ mysql/Dockerfile
@@ -1,0 +1,5 @@
+FROM mysql:5.5
+
+ENV MYSQL_DATABASE milehigh_grassroots_law
+ENV MYSQL_ROOT_PASSWORD glpdev
+COPY dump.sql /docker-entrypoint-initdb.d/
