+++ mysql/stop.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+docker stop pb-dashboard-mysql
+docker rm pb-dashboard-mysql
