+++ package.json
@@ -5,13 +5,18 @@
 	"scripts": {
 		"dev": "next",
 		"build": "next build",
-		"start": "next start"
+		"start": "next start",
+		"start-dev-db": "./mysql/start.sh",
+		"stop-dev-db": "./mysql/stop.sh"
 	},
 	"dependencies": {
+		"mysql": "^2.18.1",
 		"next": "^9.4.4",
+		"path-to-regexp": "^6.1.0",
 		"react": "^16.13.1",
 		"react-dom": "^16.13.1",
 		"react-simple-maps": "^2.1.2",
+		"sql.js": "^1.3.0",
 		"topojson": "1",
 		"victory": "^34.3.11"
 	},
