--- pages/_app.js
@@ -1,7 +1,0 @@
-import "../css/main.css";
-import React from "react";
-
-// This default export is required in a new `pages/_app.js` file.
-export default function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
-}
