+++ pages/api/[...route].js
@@ -1,0 +1,5 @@
+import handle from "api/routes/appRoutes";
+
+export default function (req, res) {
+  handle(req, res);
+}
