+++ pages/index.js
@@ -1,6 +1,6 @@
 import React from "react";
-import BrutalityByState from "../src/components/widgets/BrutalityByState";
-import BrutalityOverTime from "../src/components/widgets/BrutalityOverTime";
+import BrutalityByState from "components/widgets/BrutalityByState";
+import BrutalityOverTime from "components/widgets/BrutalityOverTime";
 
 function HomePage() {
   return (
