+++ styles/constants.js
@@ -1,0 +1,6 @@
+export const COLORS = {
+  primary: "#66fcf1",
+  text: "#0b0c10",
+  white: "#fff",
+  black: "#000",
+};
